RavenDB-13106 fix rebase issue and split tests to separate files
